requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-polymorphic==3.0.0
django-restframework==0.0.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
password-validator==1.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.2
python-http-client==3.3.1
python-magic==0.4.24
pytz==2020.4
sendgrid==6.4.7
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
whitenoise==5.2.0
